react-4 - Extracting components from main class
