+++ client/conditional-rendering/element-variables/loginButton.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+    function LoginButton(props) {
+        return (
+            <button onClick={props.onClick}>
+                Login
+            </button>
+        );
+    }
+  
+export default LoginButton;
