+++ client/conditional-rendering/element-variables/logoutButton.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+    function LogoutButton(props) {
+        return (
+            <button onClick={props.onClick}>
+                Logout
+            </button>
+        );
+    }
+
+export default LogoutButton;
